requirements.txt
Flask
Pillow
numpy
tensorflow; platform_system != "Darwin" or platform_machine != "arm64"
tensorflow-macos; platform_system == "Darwin" and platform_machine == "arm64"
tensorflow-metal; platform_system == "Darwin" and platform_machine == "arm64"
tensorflow-hub 